[pom] Add opens for java.lang
wanted this for jdk 17 only but also not sure why it was not showing up before and in general something off with argline as I cannot set it elsewhere.  Will look at that more for parent 36 changes as an outstanding issue to address.  Its probably related to osgi stuff since I've not run into it elsewhere...